Renamed Scanner.cs -> FancyScanner.cs and SimpleScanner.cs -> Scanner.cs.
Added comments to Scanner.cs.
