+++ Scantron/Scantron/Scantron.csproj
@@ -77,7 +77,7 @@
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
     <Compile Include="Question.cs" />
-    <Compile Include="Scanner.cs" />
+    <Compile Include="FancyScanner.cs" />
     <Compile Include="Scantron.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -87,7 +87,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Card.cs" />
-    <Compile Include="SimpleScanner.cs" />
+    <Compile Include="Scanner.cs" />
     <Compile Include="Student.cs" />
     <EmbeddedResource Include="Scantron.resx">
       <DependentUpon>Scantron.cs</DependentUpon>
